Remove extract 'source' package in the collector module
