+++ server/collector/src/main/java/org/bithon/server/collector/CollectorAutoConfiguration.java
@@ -18,9 +18,9 @@ package org.bithon.server.collector;
 
 import com.fasterxml.jackson.core.Version;
 import com.fasterxml.jackson.databind.Module;
-import org.bithon.server.collector.source.brpc.BrpcEventCollector;
-import org.bithon.server.collector.source.brpc.BrpcMetricCollector;
-import org.bithon.server.collector.source.brpc.BrpcTraceCollector;
+import org.bithon.server.collector.brpc.BrpcEventCollector;
+import org.bithon.server.collector.brpc.BrpcMetricCollector;
+import org.bithon.server.collector.brpc.BrpcTraceCollector;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
