+++ server/collector/src/main/java/org/bithon/server/collector/brpc/BrpcCollectorConfig.java
@@ -14,7 +14,7 @@
  *    limitations under the License.
  */
 
-package org.bithon.server.collector.source.brpc;
+package org.bithon.server.collector.brpc;
 
 import lombok.Data;
