+++ server/collector/src/main/java/org/bithon/server/collector/brpc/BrpcCollectorServer.java
@@ -14,7 +14,7 @@
  *    limitations under the License.
  */
 
-package org.bithon.server.collector.source.brpc;
+package org.bithon.server.collector.brpc;
 
 import lombok.Getter;
 import lombok.extern.slf4j.Slf4j;
