+++ server/collector/src/main/java/org/bithon/server/collector/http/TraceHttpCollector.java
@@ -14,7 +14,7 @@
  *    limitations under the License.
  */
 
-package org.bithon.server.collector.source.http;
+package org.bithon.server.collector.http;
 
 import com.fasterxml.jackson.core.JsonParseException;
 import com.fasterxml.jackson.core.JsonParser;
